Group_Management
Removed Material Design Bootstrap module and redesigned the whole website to make do with the removal of this module.

Created a variables scss file to put all style related variables such as colors, height, width and so on.

Created a button scss file to put all button related styling.

Modified the Profile service to properly take the email when modifying and edited the profile page to properly be able to modify a user's email along with his username.